Migrate interval polling to global hook and key manual updates off of global state.
